Show the native datum when the current isn't available
Previously it would think that an otherwise unspecified datum was always MLLW even if that offset wasn't available for a dataset.
